+++ tests/integration/autopilot/v1/assistant/test_dialogue.py
@@ -33,8 +33,8 @@ class DialogueTestCase(IntegrationTestCase):
             {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "assistant_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "sid": "UKkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkk",
-                "url": "https://autopilot.twilio.com/v1/Assistants/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Dialogues/UKkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkk",
+                "sid": "UKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "url": "https://autopilot.twilio.com/v1/Assistants/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Dialogues/UKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "data": {}
             }
             '''