+++ tests/integration/events/v1/test_sink.py
@@ -235,6 +235,126 @@ class SinkTestCase(IntegrationTestCase):
 
         self.assertIsNotNone(actual)
 
+    def test_read_results_in_use_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "sinks": [
+                    {
+                        "status": "initialized",
+                        "sink_configuration": {
+                            "arn": "arn:aws:kinesis:us-east-1:111111111:stream/test",
+                            "role_arn": "arn:aws:iam::111111111:role/Role",
+                            "external_id": "1234567890"
+                        },
+                        "description": "A Sink",
+                        "sid": "DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "date_created": "2015-07-30T19:00:00Z",
+                        "sink_type": "kinesis",
+                        "date_updated": "2015-07-30T19:00:00Z",
+                        "url": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "links": {
+                            "sink_test": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Test",
+                            "sink_validate": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Validate"
+                        }
+                    },
+                    {
+                        "status": "initialized",
+                        "sink_configuration": {
+                            "arn": "arn:aws:kinesis:us-east-1:222222222:stream/test",
+                            "role_arn": "arn:aws:iam::111111111:role/Role",
+                            "external_id": "1234567890"
+                        },
+                        "description": "ANOTHER Sink",
+                        "sid": "DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaab",
+                        "date_created": "2015-07-30T20:00:00Z",
+                        "sink_type": "kinesis",
+                        "date_updated": "2015-07-30T20:00:00Z",
+                        "url": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaab",
+                        "links": {
+                            "sink_test": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaab/Test",
+                            "sink_validate": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaab/Validate"
+                        }
+                    },
+                    {
+                        "status": "active",
+                        "sink_configuration": {
+                            "destination": "http://example.org/webhook",
+                            "method": "POST",
+                            "batch_events": true
+                        },
+                        "description": "A webhook Sink",
+                        "sid": "DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac",
+                        "date_created": "2015-07-30T21:00:00Z",
+                        "sink_type": "webhook",
+                        "date_updated": "2015-07-30T21:00:00Z",
+                        "url": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac",
+                        "links": {
+                            "sink_test": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac/Test",
+                            "sink_validate": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac/Validate"
+                        }
+                    }
+                ],
+                "meta": {
+                    "page": 0,
+                    "page_size": 20,
+                    "first_page_url": "https://events.twilio.com/v1/Sinks?InUse=True&PageSize=20&Page=0",
+                    "previous_page_url": null,
+                    "url": "https://events.twilio.com/v1/Sinks?InUse=True&PageSize=20&Page=0",
+                    "next_page_url": null,
+                    "key": "sinks"
+                }
+            }
+            '''
+        ))
+
+        actual = self.client.events.v1.sinks.list()
+
+        self.assertIsNotNone(actual)
+
+    def test_read_results_status_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "sinks": [
+                    {
+                        "status": "active",
+                        "sink_configuration": {
+                            "destination": "http://example.org/webhook",
+                            "method": "POST",
+                            "batch_events": true
+                        },
+                        "description": "A webhook Sink",
+                        "sid": "DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac",
+                        "date_created": "2015-07-30T21:00:00Z",
+                        "sink_type": "webhook",
+                        "date_updated": "2015-07-30T21:00:00Z",
+                        "url": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac",
+                        "links": {
+                            "sink_test": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac/Test",
+                            "sink_validate": "https://events.twilio.com/v1/Sinks/DGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaac/Validate"
+                        }
+                    }
+                ],
+                "meta": {
+                    "page": 0,
+                    "page_size": 20,
+                    "first_page_url": "https://events.twilio.com/v1/Sinks?Status=active&PageSize=20&Page=0",
+                    "previous_page_url": null,
+                    "url": "https://events.twilio.com/v1/Sinks?Status=active&PageSize=20&Page=0",
+                    "next_page_url": null,
+                    "key": "sinks"
+                }
+            }
+            '''
+        ))
+
+        actual = self.client.events.v1.sinks.list()
+
+        self.assertIsNotNone(actual)
+
     def test_update_request(self):
         self.holodeck.mock(Response(500, ''))