+++ tests/integration/insights/v1/room/test_participant.py
@@ -32,7 +32,7 @@ class ParticipantTestCase(IntegrationTestCase):
             '''
             {
                 "publisher_info": {},
-                "edge_location": "Ashburn",
+                "edge_location": "ashburn",
                 "join_time": "2015-07-30T20:00:00Z",
                 "leave_time": "2015-07-30T20:00:00Z",
                 "end_reason": "disconnected_via_api",
@@ -112,7 +112,7 @@ class ParticipantTestCase(IntegrationTestCase):
                 "participants": [
                     {
                         "publisher_info": {},
-                        "edge_location": "Ashburn",
+                        "edge_location": "ashburn",
                         "join_time": "2015-07-30T20:00:00Z",
                         "leave_time": "2015-07-30T20:00:00Z",
                         "end_reason": "disconnected_via_api",