+++ tests/integration/messaging/v1/service/test_phone_number.py
@@ -167,7 +167,7 @@ class PhoneNumberTestCase(IntegrationTestCase):
             200,
             '''
             {
-                "sid": "SCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "sid": "PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "service_sid": "MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2015-07-30T20:12:31Z",
@@ -175,7 +175,7 @@ class PhoneNumberTestCase(IntegrationTestCase):
                 "phone_number": "12345",
                 "country_code": "US",
                 "capabilities": [],
-                "url": "https://messaging.twilio.com/v1/Services/MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/PhoneNumbers/SCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "url": "https://messaging.twilio.com/v1/Services/MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/PhoneNumbers/PNaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
         ))