+++ tests/integration/messaging/v1/test_external_campaign.py
@@ -33,6 +33,7 @@ class ExternalCampaignTestCase(IntegrationTestCase):
             201,
             '''
             {
+                "sid": "QE2c6890da8086d771620e9b13fadeba0b",
                 "messaging_service_sid": "MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "campaign_id": "CMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",