+++ tests/integration/monitor/v1/test_alert.py
@@ -42,7 +42,7 @@ class AlertTestCase(IntegrationTestCase):
                 "request_method": "GET",
                 "request_url": "http://www.example.com",
                 "request_variables": "request_variables",
-                "resource_sid": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "resource_sid": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "response_body": "response_body",
                 "response_headers": "response_headers",
                 "request_headers": "request_headers",
@@ -109,7 +109,7 @@ class AlertTestCase(IntegrationTestCase):
                         "more_info": "more_info",
                         "request_method": "GET",
                         "request_url": "http://www.example.com",
-                        "resource_sid": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "resource_sid": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "sid": "NOaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "url": "https://monitor.twilio.com/v1/Alerts/NOaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "service_sid": "PNe2cd757cd5257b0217a447933a0290d2"