+++ tests/integration/wireless/v1/test_command.py
@@ -156,7 +156,7 @@ class CommandTestCase(IntegrationTestCase):
                     {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "command": "command",
-                        "command_mode": "not_confirmable",
+                        "command_mode": "binary",
                         "date_created": "2015-07-30T20:00:00Z",
                         "date_updated": "2015-07-30T20:00:00Z",
                         "delivery_receipt_requested": true,