+++ twilio/rest/events/v1/sink/__init__.py
@@ -40,7 +40,7 @@ class SinkList(ListResource):
         """
         Create the SinkInstance
 
-        :param unicode description: Sink Description
+        :param unicode description: Sink Description.
         :param dict sink_configuration: JSON Sink configuration.
         :param SinkInstance.SinkType sink_type: Sink type.
 
@@ -57,13 +57,16 @@ class SinkList(ListResource):
 
         return SinkInstance(self._version, payload, )
 
-    def stream(self, limit=None, page_size=None):
+    def stream(self, in_use=values.unset, status=values.unset, limit=None,
+               page_size=None):
         """
         Streams SinkInstance records from the API as a generator stream.
         This operation lazily loads records as efficiently as possible until the limit
         is reached.
         The results are returned as a generator, so this operation is memory efficient.
 
+        :param bool in_use: A boolean to return sinks used/not used by a subscription.
+        :param unicode status: A string to filter sinks by status.
         :param int limit: Upper limit for the number of records to return. stream()
                           guarantees to never return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -76,16 +79,19 @@ class SinkList(ListResource):
         """
         limits = self._version.read_limits(limit, page_size)
 
-        page = self.page(page_size=limits['page_size'], )
+        page = self.page(in_use=in_use, status=status, page_size=limits['page_size'], )
 
         return self._version.stream(page, limits['limit'])
 
-    def list(self, limit=None, page_size=None):
+    def list(self, in_use=values.unset, status=values.unset, limit=None,
+             page_size=None):
         """
         Lists SinkInstance records from the API as a list.
         Unlike stream(), this operation is eager and will load `limit` records into
         memory before returning.
 
+        :param bool in_use: A boolean to return sinks used/not used by a subscription.
+        :param unicode status: A string to filter sinks by status.
         :param int limit: Upper limit for the number of records to return. list() guarantees
                           never to return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -96,14 +102,17 @@ class SinkList(ListResource):
         :returns: Generator that will yield up to limit results
         :rtype: list[twilio.rest.events.v1.sink.SinkInstance]
         """
-        return list(self.stream(limit=limit, page_size=page_size, ))
+        return list(self.stream(in_use=in_use, status=status, limit=limit, page_size=page_size, ))
 
-    def page(self, page_token=values.unset, page_number=values.unset,
+    def page(self, in_use=values.unset, status=values.unset,
+             page_token=values.unset, page_number=values.unset,
              page_size=values.unset):
         """
         Retrieve a single page of SinkInstance records from the API.
         Request is executed immediately
 
+        :param bool in_use: A boolean to return sinks used/not used by a subscription.
+        :param unicode status: A string to filter sinks by status.
         :param str page_token: PageToken provided by the API
         :param int page_number: Page Number, this value is simply for client state
         :param int page_size: Number of records to return, defaults to 50
@@ -111,7 +120,13 @@ class SinkList(ListResource):
         :returns: Page of SinkInstance
         :rtype: twilio.rest.events.v1.sink.SinkPage
         """
-        data = values.of({'PageToken': page_token, 'Page': page_number, 'PageSize': page_size, })
+        data = values.of({
+            'InUse': in_use,
+            'Status': status,
+            'PageToken': page_token,
+            'Page': page_number,
+            'PageSize': page_size,
+        })
 
         response = self._version.page(method='GET', uri=self._uri, params=data, )