+++ twilio/rest/video/v1/composition/__init__.py
@@ -16,9 +16,6 @@ from twilio.base.page import Page
 
 
 class CompositionList(ListResource):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     def __init__(self, version):
         """
@@ -221,9 +218,6 @@ class CompositionList(ListResource):
 
 
 class CompositionPage(Page):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     def __init__(self, version, response, solution):
         """
@@ -262,9 +256,6 @@ class CompositionPage(Page):
 
 
 class CompositionContext(InstanceContext):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     def __init__(self, version, sid):
         """
@@ -314,9 +305,6 @@ class CompositionContext(InstanceContext):
 
 
 class CompositionInstance(InstanceResource):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     class Status(object):
         ENQUEUED = "enqueued"