+++ twilio/rest/video/v1/composition_hook.py
@@ -16,9 +16,6 @@ from twilio.base.page import Page
 
 
 class CompositionHookList(ListResource):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     def __init__(self, version):
         """
@@ -223,9 +220,6 @@ class CompositionHookList(ListResource):
 
 
 class CompositionHookPage(Page):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     def __init__(self, version, response, solution):
         """
@@ -264,9 +258,6 @@ class CompositionHookPage(Page):
 
 
 class CompositionHookContext(InstanceContext):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     def __init__(self, version, sid):
         """
@@ -354,9 +345,6 @@ class CompositionHookContext(InstanceContext):
 
 
 class CompositionHookInstance(InstanceResource):
-    """ PLEASE NOTE that this class contains preview products that are subject
-    to change. Use them with caution. If you currently do not have developer
-    preview access, please contact help@twilio.com. """
 
     class Format(object):
         MP4 = "mp4"